worker/requirements.txt
openai>=1.60.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
requests>=2.31.0
Pillow>=10.0.0
